Fixes display issues with lists
When I had a link and content on the same bullet, a space would be
introduced in the rendered HTML (a liquid issue apparently). That space
would be injected and would always display causing weird visual
indentation. Switching to an image instead of the li:before fixes that.
Not sure why but it works.

Fix taken from pages-themes/hacker#5.